requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
prometheus-client==0.21.1
pytest==8.3.4
httpx==0.28.1
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi
